api/pyproject.toml
[project]
name = "learnwithai-api"
version = "0.1.0"
description = "FastAPI adapter for LearnWithAI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1",
  "uvicorn>=0.30,<1",
  "websockets>=15.0,<16",
  "pyjwt>=2.12.1",
  "python-multipart>=0.0.12",
  "learnwithai-core",
  "learnwithai-jobqueue",
  "httpx>=0.28.1",
  "aio-pika>=9.5,<10",
]

[tool.uv.sources]
learnwithai-core = { workspace = true }
learnwithai-jobqueue = { workspace = true }

[dependency-groups]
dev = [
  "pytest>=8.3,<10",
  "pytest-cov>=6.0,<7",
  "httpx>=0.28,<1",
  "ruff>=0.11,<0.16",
  "pyright>=1.1.400,<2",
]

[tool.uv.build-backend]
module-name = "api"

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"
